Move configure files into config directory
